Added Login
Added admin login functionality.
Protected  admin pages by requiring login.
Implemented extensible user account framework in case future  CS students want to add the ability to have multiple accounts.
Added a message on the navbar that displays a welcome message and a link to log out when the user is logged in and a log in link when the user is logged out.
Changed "Nerdboard" to "Comp Sci Corner" in the page titles.
Gave each page a custom title.